Adds pop up when you click on cards
makes all card images buttons with an onclick function that passes in the index and imageurl. When you click on a card has a pop up with the correct card you clicked on and buttons to should where to store the card. The goal of this is to possibly store the index of the card to the database with the user.