moved all javascript to one file
I moved all javascript files to one file. it uses if statements to determine which page you are on. Also, put logout button on all the pages and when you do log out it redirects you to the index or home page. The goal is to get it to where if you aren't logged in you cannot access any other page other than the home page.