GetCardIndex function
The line added to cardDecide gave the new attribute 'index' so that we can access the current card index.

The new function GetCardIndex is used as a test of the attribute and I put in on the button  "Add to Trunk" to make sure it works with all cards. This could help us set up the users database like Austin had in mind with saving indexs of cards to the users.