new document in database when new user signs in
when a new user signs in,  it creates a new document with the fields trunk, deck1, deck2, deck3, and desired. I'm still working on being able to click on a card and store it in a field which is the big chunk of code I have commented out. I also made it so that if someone is signed out they can't type in http://localhost:5000/decks.html, trunk.html, or desired.html and go to those pages. So, the get redirected to the home page.