Removing cards
Added in Removing cards from the different pages. Also, you can click on cards and it will pop up with options on all pages. When you click the button, it reloads the page after a couple seconds to make sure the card index is removed from the database and to show the updated contents of the page.  Also, added that you can only have 60 cards in a deck.